Enable pKEK rewrap with SimpleCrypto
This patch adds a new sub-command to the `barbican-manage` command line
tool to rewrap all pKEKs using the currently active KEK from the config
file.
